refactor: fix remaining import paths after audit module restructure
